fulfilling part of the recommendations for the first task
